Add startup welcome and version components for iOS and Android; remove default message components
