+++ frontend/app/(views)/app-navegante-android/startup/20250115/page.tsx
@@ -1,0 +1,9 @@
+/* * */
+
+import { AppNaveganteStartup20250115 } from '@/components/app-navegante/startup/AppNaveganteStartup20250115';
+
+/* * */
+
+export default function Page() {
+	return <AppNaveganteStartup20250115 />;
+}
