+++ frontend/app/(views)/app-navegante-android/startup/welcome/page.tsx
@@ -1,0 +1,9 @@
+/* * */
+
+import { AppNaveganteStartupWelcome } from '@/components/app-navegante/startup/AppNaveganteStartupWelcome';
+
+/* * */
+
+export default function Page() {
+	return <AppNaveganteStartupWelcome />;
+}
