+++ frontend/app/(views)/app-navegante-ios/startup/20250115/page.tsx
@@ -1,0 +1,9 @@
+/* * */
+
+import { AppNaveganteStartup20250115 } from '@/components/app-navegante/startup/AppNaveganteStartup20250115';
+
+/* * */
+
+export default function Page() {
+	return <AppNaveganteStartup20250115 />;
+}
