--- frontend/app/(views)/app-navegante-ios/startup/[anything]/default/page.tsx
@@ -1,9 +1,0 @@
-/* * */
-
-import { AppNaveganteStartupMessageDefault } from '@/components/app-navegante/AppNaveganteStartupMessageDefault';
-
-/* * */
-
-export default function Page() {
-	return <AppNaveganteStartupMessageDefault />;
-}
