+++ frontend/app/(views)/app-navegante-ios/startup/welcome/page.tsx
@@ -1,0 +1,9 @@
+/* * */
+
+import { AppNaveganteStartupWelcome } from '@/components/app-navegante/startup/AppNaveganteStartupWelcome';
+
+/* * */
+
+export default function Page() {
+	return <AppNaveganteStartupWelcome />;
+}
