--- frontend/components/app-navegante/AppNaveganteStartupMessageDefault/styles.module.css
@@ -1,55 +1,0 @@
-/* * */
-/* CONTAINER & WRAPPERS */
-
-.container,
-.introWrapper,
-.actionsWrapper {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	text-align: center;
-}
-
-/* * */
-/* INTRO */
-
-.introWrapper {
-	gap: var(--size-spacing-15);
-	padding: var(--size-spacing-30) var(--size-spacing-15);
-	background-color: var(--color-system-background-100);
-}
-
-.icon {
-	display: flex;
-    width: 60px;
-	height: 60px;
-	color: var(--color-status-ok-text);
-}
-
-.heading {
-	font-size: 26px;
-	font-weight: var(--font-weight-bold);
-	color: var(--color-system-text-100);
-}
-
-.subheading {
-	font-size: 18px;
-	font-weight: var(--font-weight-semibold);
-	color: var(--color-system-text-200);
-}
-
-.text {
-	font-size: 16px;
-	font-weight: var(--font-weight-semibold);
-	color: var(--color-system-text-300);
-}
-
-/* * */
-/* ACTIONS */
-
-.actionsWrapper {
-	gap: var(--size-spacing-30);
-	padding: var(--size-spacing-30) var(--size-spacing-15);
-}
