+++ frontend/themes/app-navegante/overrides/Button.module.css
@@ -4,7 +4,7 @@
 .root,
 .root.variantPrimary {
 	width: 100%;
-	height: 60px;
+	height: 50px;
 	color: var(--color-primary-black);
     background-color: var(--color-brand);
     border-radius: 999px;
@@ -16,8 +16,8 @@
 }
 
 .root.variantSecondary {
-	color: var(--color-system-text-900);
-	background-color: var(--color-system-background-900);
+	color: var(--color-system-background-100);
+	background-color: var(--color-system-text-100);
 }
 
 .root.variantMuted {
